Renamed vl1l2_statiatics.py to be vl1l2_statistics.py
