+++ metcalcpy/sum_stat.py
@@ -45,7 +45,7 @@ from metcalcpy.util.grad_statistics import *
 from metcalcpy.util.sl1l2_statistics import *
 from metcalcpy.util.val1l2_statistics import *
 from metcalcpy.util.vcnt_statistics import *
-from metcalcpy.util.vl1l2_statiatics import *
+from metcalcpy.util.vl1l2_statistics import *
 from metcalcpy.util.pstd_statistics import *
 from metcalcpy.util.nbrcnt_statistics import *
 from metcalcpy.util.ecnt_statistics import *
